+++ docs/plugins/parameters.rst
@@ -2,6 +2,4 @@
 Loader: Parameterized Tests
 ===========================
 
-.. autofunction :: nose2.tools.params
-
 .. autoplugin :: nose2.plugins.loader.parameters.Parameters
