+++ docs/tools.rst
@@ -8,4 +8,5 @@ Tools for Test Authors
 .. toctree ::
    :maxdepth: 2
 
+   params
    such_dsl
